Migrated db test to helpers.test.js
* Added basename utility test
* Confirmed that vanilla db.js writes as expected
