+++ src/content.js
@@ -3,7 +3,7 @@
  * @Date:   2018-01-02T09:34:17-08:00
  * @Email:  alec@bubblegum.academy
  * @Last modified by:   alechp
- * @Last modified time: 2018-01-11T16:18:06-08:00
+ * @Last modified time: 2018-01-12T11:15:49-08:00
  */
 
 const fs = require("fs-extra");
