+++ tests/file.test.js
@@ -3,7 +3,7 @@
  * @Date:   2018-01-04T12:43:32-08:00
  * @Email:  alec@bubblegum.academy
  * @Last modified by:   alechp
- * @Last modified time: 2018-01-12T11:13:49-08:00
+ * @Last modified time: 2018-01-12T11:42:26-08:00
  */
 
 const log = console.log;
@@ -48,5 +48,3 @@ afterAll(() => {
       log(`${chalk.yellow("file.test.js")} failed cleanup. ${chalk.red(err)}`);
     });
 });
-
-// TODO: Setup TRAVIS-CI for test suite
