+++ tests/sandbox.history.json
@@ -1,6 +1,6 @@
 {
   "tests": {
-    "lastTestRan": "Jan, 12th, 2018, 11:15:1741",
+    "lastTestRan": "Jan, 12th, 2018, 11:50:0527",
     "lastPlainFileCreated": "/Users/alechp/Code/servexyz/generator-generator/lib/tests/.sandbox/foo.bar",
     "lastSymlinkFileCreated": "/Users/alechp/Code/servexyz/generator-generator/lib/tests/.sandbox/.gitignore"
   }
